move admin pilots routes to new controller
