+++ dashboard/app/views/admin_pilots/pilots.html.haml
