+++ dashboard/app/views/admin_pilots/show_pilot.html.haml
@@ -8,7 +8,7 @@
 %p Separate emails by line. For example:
 %p teacher@school.org<br>teacher2@school.org<br>teacher3@school.org
 %p Note: Use the lower right-hand corner to drag the box to any size.
-= form_tag controller: 'admin_search', action: 'add_to_pilot', method: "post" do
+= form_tag controller: 'admin_pilots', action: 'add_to_pilot', method: "post" do
   = text_area_tag :email, nil, width: '300px', height: '300px'
   = hidden_field_tag 'pilot_name', @pilot_name
   = submit_tag "Add", class: 'btn'
@@ -23,7 +23,7 @@
       %td.email
         %span= email
       %td.actions
-        = form_tag({controller: 'admin_search', action: 'remove_from_pilot', method: "post"}, style: 'margin: 5px;')do
+        = form_tag({controller: 'admin_pilots', action: 'remove_from_pilot', method: "post"}, style: 'margin: 5px;')do
           = hidden_field_tag 'email', email
           = hidden_field_tag 'pilot_name', @pilot_name
           = submit_tag "Remove", class: 'btn-primary', data:{confirm: "You are about to remove " + email + " from this pilot.  Once you do this all progress this participant has made will be lost.  Are you sure?"}
