+++ dashboard/config/routes.rb
@@ -613,11 +613,11 @@ Dashboard::Application.routes.draw do
     get '/admin/lookup_section', to: 'admin_search#lookup_section', as: 'lookup_section'
     post '/admin/lookup_section', to: 'admin_search#lookup_section'
     post '/admin/undelete_section', to: 'admin_search#undelete_section', as: 'undelete_section'
-    get '/admin/pilots/', to: 'admin_search#pilots', as: 'pilots'
-    post '/admin/pilots/', to: 'admin_search#create_pilot', as: 'create_pilot'
-    get '/admin/pilots/:pilot_name', to: 'admin_search#show_pilot', as: 'show_pilot'
-    post '/admin/add_to_pilot', to: 'admin_search#add_to_pilot', as: 'add_to_pilot'
-    post '/admin/remove_from_pilot', to: 'admin_search#remove_from_pilot', as: 'remove_from_pilot'
+    get '/admin/pilots/', to: 'admin_pilots#pilots', as: 'pilots'
+    post '/admin/pilots/', to: 'admin_pilots#create_pilot', as: 'create_pilot'
+    get '/admin/pilots/:pilot_name', to: 'admin_pilots#show_pilot', as: 'show_pilot'
+    post '/admin/add_to_pilot', to: 'admin_pilots#add_to_pilot', as: 'add_to_pilot'
+    post '/admin/remove_from_pilot', to: 'admin_pilots#remove_from_pilot', as: 'remove_from_pilot'
 
     # internal engineering dashboards
     get '/admin/dynamic_config', to: 'dynamic_config#show', as: 'dynamic_config_state'
